feat: automatic escrow deposits
